Read test type from file
This changes the tx_script to read from test-type.txt to get a string
indicating the type of test that is being performed. The test type is
inserted into the capture filename and stored in the capture file.

The process script reads the test type and uses it for the processed
data. This is to allow for some indication of the type of test that is
being performed when the data is analyzed later