Fixed bug in time_decode
The IRIG format only gives day of year and not month or day of month. To
get the correct month and day of month a date vector is generated with a
dummy month and day of year. datetime will wrap the day of year to the
correct month. This dummy month was prevously 0 when it should be 1
causing the decoded month to be off.