Update the way periodic checks are done
This changes when periodic checks happen. Now the first trial is always
checked and if there are less than 10 trials there are 2 checks
regardless of how many trials there are (unless there is only one
trial). This helps with detecting problems as they occur so that they
can be fixed faster and less time is lost.